record audit refactored and bug fixed on downlowding state on record audit
