+++ packages/client/src/App.tsx
@@ -56,7 +56,7 @@ import { Switch } from 'react-router'
 import { AppStore } from './store'
 import { CorrectionForm, CorrectionReviewForm } from './views/CorrectionForm'
 import { VerifyCorrector } from './views/CorrectionForm/VerifyCorrector'
-import { RecordAudit } from './views/Home/RecordAudit'
+import { RecordAudit } from './views/RecordAudit/RecordAudit'
 import { ChangePhonePage } from './views/Settings/ChangePhonePage'
 import { ApplicationConfig } from './views/SysAdmin/Config/Application'
 import { CertificatesConfig } from './views/SysAdmin/Config/Certificates'
